requirements.txt
prefect==3.6.2
requests==2.32.4
urllib3==2.0.7
python-dateutil==2.8.2
pytest==7.4.0
google-generativeai==0.8.5
google-genai>=1.0.0
python-dotenv==0.21.0
openai==2.11.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform>=1.49.0
google-adk>=0.3.0
